[svn] Replaced JFlex/CUP-based parser with one based on JavaCC. This works around a limit in CUP's parser table size.
This helped on a few other fronts:
- no more duplicated parsers (A1, R1C1, Rewrite)
- one less runtime dependency (CUP runtime)
- moved the OpenFormula/Excel-conformant parser out of the excel package into the compiler core

The hacking and design docs, however, are not updated yet. I shall do this soon.

I attempted to fix a problem with the TODAY test, but it is still not entirely satisfactory. Vladimir is going to address the problems there at the root once he gets around to it.